Split CMakeLists.txt into a project of three files
and move C++ SDK definitions to a .cpp file
