+++ rerun_cpp/CMakeLists.txt
@@ -1,0 +1,6 @@
+cmake_minimum_required(VERSION 3.16)
+
+project(rerun_cpp_proj LANGUAGES CXX)
+
+add_subdirectory(src) # The Rerun C++ SDK library
+add_subdirectory(example)
