--- rerun_cpp/example/.gitignore
@@ -1,3 +1,0 @@
-*.bin
-*.o
-build/
