+++ rerun_cpp/src/CMakeLists.txt
@@ -1,0 +1,58 @@
+cmake_minimum_required(VERSION 3.16)
+
+add_library(rerun_sdk rerun.cpp)
+
+# ------------------------------------------------------------------------------
+
+# For rerun.h (Rerun C SDK):
+include_directories(SYSTEM ${CMAKE_CURRENT_SOURCE_DIR}/../../crates/rerun_c/src)
+
+# Make sure the compiler can find include files for rerun
+# when other libraries or executables link to rerun:
+target_include_directories(rerun_sdk PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
+
+# ------------------------------------------------------------------------------
+# Loguru logging library (https://github.com/emilk/loguru):
+set(CMAKE_DL_LIBS "dl") # Required by Loguru for backtraces
+
+# Loguru, see https://github.com/emilk/loguru/blob/master/loguru_cmake_example/CMakeLists.txt
+include(FetchContent)
+FetchContent_Declare(LoguruGitRepo
+    GIT_REPOSITORY "https://github.com/emilk/loguru" # can be a filesystem path
+    GIT_TAG "master"
+)
+
+# set any loguru compile-time flags before calling MakeAvailable()
+set(LOGURU_STACKTRACES 1)
+FetchContent_MakeAvailable(LoguruGitRepo) # defines target 'loguru::loguru'
+
+target_link_libraries(rerun_sdk PRIVATE loguru::loguru)
+
+# ------------------------------------------------------------------------------
+if(APPLE)
+    target_link_libraries(rerun_sdk PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/../../target/debug/librerun_c.a)
+    target_link_libraries(rerun_sdk PRIVATE "-framework CoreFoundation" "-framework IOKit")
+endif()
+
+if(LINUX)
+    target_link_libraries(rerun_sdk PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/../../target/debug/librerun_c.a)
+    target_link_libraries(rerun_sdk PRIVATE "-lm")
+endif()
+
+# -----------------------------------------------------------------------------
+# Arrow:
+option(ARROW_LINK_SHARED "Link to the Arrow shared library" ON)
+
+find_package(Arrow REQUIRED)
+
+# Arrow requires a C++17 compliant compiler
+set(CMAKE_CXX_STANDARD_REQUIRED ON)
+
+message(STATUS "Arrow version: ${ARROW_VERSION}")
+message(STATUS "Arrow SO version: ${ARROW_FULL_SO_VERSION}")
+
+if(ARROW_LINK_SHARED)
+    target_link_libraries(rerun_sdk PRIVATE Arrow::arrow_shared)
+else()
+    target_link_libraries(rerun_sdk PRIVATE Arrow::arrow_static)
+endif()
